[add] 石をランダムで生成するクラスを追加
